сделан viewModel
